// src/app/models/comunidades.model.ts
export class ComunidadesModel {
    Id: number;
    NombreComunidad: string;
    Direccion: string;
    NombreContacto: string;
    FonoContacto: string;
    EmailContacto: string;
    RutaFirebase: string;
    MaximoUsuario: number;
    MaximoPropiedades: number;
    Habilitado: string;
    Administrador1: string;
    Administrador2: string;
    // EstadoPorton: string;
}